pipeline-func/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.20.0
pandas==2.2.2
